Fixing missing vendor.* files in Docker image

+++ docker_entrypoint.sh
@@ -8,6 +8,13 @@ fi
 if [ -n "${PROXY_TIMEOUT_IN_SECS:-}" ]; then
   export PROXY_TIMEOUT_DIRECTIVES="proxy_connect_timeout ${PROXY_TIMEOUT_IN_SECS}s; proxy_send_timeout ${PROXY_TIMEOUT_IN_SECS}s; proxy_read_timeout ${PROXY_TIMEOUT_IN_SECS}s; send_timeout ${PROXY_TIMEOUT_IN_SECS}s;"
 fi
+
+# Ensure the CSS & JS dependencies files are present:
+if ! [ -s /usr/share/nginx/html/vendor.css ] || ! [ -s /usr/share/nginx/html/vendor.js ]; then
+    echo 'At least one of vendor.css / vendor.js are missing or empty' >&2
+    exit 1
+fi
+
 # Inject env variables in hesperides.conf:
 envsubst </etc/nginx/conf.d/hesperides.conf.template >/etc/nginx/conf.d/hesperides.conf
